html/semantics/document-metadata/styling/LinkStyle.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: Link_Style</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/document-metadata.html#styling">
    <link id="style1" rel="text" title="Intel" href="./support/unmatch.css">
    <link id="style2" rel="alternate stylesheet" type="text/css" title="" href="./support/emptytitle.css">
    <link id="style3" rel="alternate stylesheet" type="text/css" href="./support/notitle.css">
    <link id="style5" rel="stylesheet" type="text/css" href="./support/normal.css">
    <link id="style6" rel="alternate stylesheet" type="text/css" href="./support/normal.css" title="./support/alternate.css">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style id="style4" type="text/html">
      #test {
        height:100px;
        width:100px;
      }
    </style>
    <style id="style7" type="text/css" media="all" title="./support/alternate.css">
      #test {
        background-color:green;
      }
    </style>
  </head>
  <body>
    <div id="log"></div>
    <div id="test" style="display:none">STYLING TEST</div>

    <script>
      var styles = [];
      test(function() {
        for (var i = 0; i < 4; i++) {
          styles[i] = document.getElementById("style" + (i + 1));
          assert_equals(styles[i].sheet, null, "The 'linkStyle' of style" + (i + 1) + " should be null.");
          assert_false(styles[i].disabled, "The disabled attribute of style" + (i + 1) + " should be fasle");
        }
      }, "Link_Style_return_null");

      test(function() {
        var styles = [];
        for (var i = 4; i < 7; i++) {
          styles[i] = document.getElementById("style" + (i + 1));
          assert_true(styles[i].sheet instanceof StyleSheet, "The linkStyle of style" + (i + 1) + " should be a StyleSheet object.");
          assert_equals(styles[i].disabled, styles[i].sheet.disabled, "The disabled attribute" + (i + 1) + " should equal to the same attribute of StyleSheet.");
        }
      }, "Link_Style_return_StyleSheet");

      test(function() {
        var style2 = document.getElementById("style2");
        var style5 = document.getElementById("style5");
        var style6 = document.getElementById("style6");
        var style6 = document.getElementById("style7");
        assert_equals(style2.title, "", "The title attribute of style2 is incorrect.");
        assert_equals(style5.title, "", "The title attribute of style5 is incorrect.");
        assert_equals(style6.title, "./support/alternate.css", "The title attribute is incorrect.");
        assert_equals(style7.title, "./support/alternate.css", "The title attribute is incorrect.");
      }, "Link_Style_title");

      test(function() {
        var style5 = document.getElementById("style5");
        var style7 = document.getElementById("style7");
        assert_equals(style5.media, "", "The media attribute of style5 is incorrect.");
        assert_equals(style7.media, "all", "The title attribute is incorrect.");
      }, "Link_Style_media");
    </script>
  </body>
</html>
